test(amplify-e2e-tests): code review suggestions
code review suggestions: verbose error handling, comments, removal of unneeded tsline disabling and
test assertions, moving helper functions to utility file
