+++ packages/amplify-e2e-tests/__tests__/auth.test.ts
@@ -5,11 +5,7 @@ import {
   amplifyPushAuth
 } from '../src/init';
 import { addAuthWithDefault, addAuthWithDefaultSocial } from '../src/categories/auth';
-import { createNewProjectDir, deleteProjectDir, getProjectMeta } from '../src/utils';
-import * as AWS from 'aws-sdk';
-
-AWS.config.update({ region: 'us-west-2' });
-const CognitoIdentityServiceProvider = AWS.CognitoIdentityServiceProvider;
+import { createNewProjectDir, deleteProjectDir, getProjectMeta, getUserPool, getUserPoolClients } from '../src/utils';
 
 describe('amplify add auth', () => {
   let projRoot: string;
@@ -42,39 +38,10 @@ describe('amplify add auth', () => {
     const userPool = await getUserPool(id);
     const clients = await getUserPoolClients(id);
     await expect(userPool.UserPool).toBeDefined();
-    await expect(clients).toBeDefined();
-    await expect(clients.length).toEqual(2);
+    await expect(clients).toHaveLength(2);
     await expect(clients[0].UserPoolClient.CallbackURLs[0]).toEqual('https://www.google.com/');
     await expect(clients[0].UserPoolClient.LogoutURLs[0]).toEqual('https://www.nytimes.com/');
-    await expect(clients[0].UserPoolClient.SupportedIdentityProviders.length).toEqual(4);
+    await expect(clients[0].UserPoolClient.SupportedIdentityProviders).toHaveLength(4);
 
   });
 });
-
-const getUserPool = async (userpoolId) => {
-  let res;
-  try {
-    res = await new CognitoIdentityServiceProvider().describeUserPool({UserPoolId: userpoolId}).promise();
-  } catch (e) {
-    console.log(e);
-  }
-  return res;
-}
-
-const getUserPoolClients = async (userpoolId) => {
-  const provider = new CognitoIdentityServiceProvider();
-  let res = [];
-  try {
-    let clients = await provider.listUserPoolClients({UserPoolId: userpoolId}).promise();
-    for (let i = 0; i < clients.UserPoolClients.length; i++) {
-      let clientData = await provider.describeUserPoolClient({
-        UserPoolId: userpoolId,
-        ClientId: clients.UserPoolClients[i].ClientId
-      }).promise();
-      res.push(clientData);
-    }
-  } catch (e) {
-    console.log(e);
-  }
-  return res;
-}
