+++ packages/amplify-e2e-tests/sample.env
@@ -1,0 +1,12 @@
+AWS_ACCESS_KEY_ID=<your-access-key>
+
+AWS_SECRET_ACCESS_KEY=<your-secret-access-key>
+
+FACEBOOK_APP_ID=<your-facebook-app-id>
+FACEBOOK_APP_SECRET=<your-facebook-app-secret>
+
+GOOGLE_APP_ID=<your-google-app-id>
+GOOGLE_APP_SECRET=<your-google-app-secret>
+
+AMAZON_APP_ID=<your-amazon-app-id>
+AMAZON_APP_SECRET=<your-amazon-app-secret>
